requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
colorama==0.4.3
Django==3.1
isort==5.4.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pylint==2.6.0
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
wrapt==1.12.1